Extract WorkTime calendar into dedicated component
